+++ app/wyzebridge/mqtt.py
@@ -26,10 +26,13 @@ def mqtt_enabled(func):
         try:
             return func(*args, **kwargs)
         except ConnectionRefusedError:
-            logger.warning("[MQTT] connection refused. Disabling MQTT.")
+            logger.error("[MQTT] connection refused. Disabling MQTT.")
+            MQTT_ENABLED = False
+        except TimeoutError:
+            logger.error("[MQTT] TimeoutError. Disabling MQTT.")
             MQTT_ENABLED = False
         except Exception as ex:
-            logger.warning(f"[MQTT] {ex}")
+            logger.error(f"[MQTT] {ex}")
 
     return wrapper
 
@@ -39,7 +42,6 @@ def wyze_discovery(cam: WyzeCamera, cam_uri: str) -> None:
     """Add Wyze camera to MQTT if enabled."""
     base = f"wyzebridge/{cam_uri or cam.name_uri}/"
     msgs = [(f"{base}state", "stopped")]
-
     if MQTT_DISCOVERY:
         base_payload = {
             "device": {
@@ -53,61 +55,19 @@ def wyze_discovery(cam: WyzeCamera, cam_uri: str) -> None:
             },
         }
 
-        entities = {
-            "preview": {
-                "availability_topic": f"{base}state",
-                "payload_not_available": "stopped",
-                "type": "camera",
-                "payload": {
-                    "topic": f"{base}image",
-                    "icon": "mdi:cctv",
-                },
-            },
-            "ir": {
-                "type": "switch",
-                "payload": {
-                    "state_topic": f"{base}irled",
-                    "command_topic": f"{base}irled/set",
-                    "payload_on": 1,
-                    "payload_off": 2,
-                    "icon": "mdi:lightbulb-night",
-                },
-            },
-            "night_vision": {
-                "type": "switch",
-                "payload": {
-                    "state_topic": f"{base}night_vision",
-                    "command_topic": f"{base}night_vision/set",
-                    "payload_on": 3,
-                    "payload_off": 2,
-                    "icon": "mdi:weather-night",
-                },
-            },
-            "signal": {
-                "type": "sensor",
-                "payload": {
-                    "state_topic": f"{base}wifi",
-                    "icon": "mdi:wifi",
-                    "entity_category": "diagnostic",
-                },
-            },
-            "audio": {
-                "type": "sensor",
-                "payload": {
-                    "state_topic": f"{base}audio",
-                    "icon": "mdi:volume-high",
-                    "entity_category": "diagnostic",
-                },
-            },
-        }
-        for entity, data in entities.items():
+        for entity, data in get_entities(base).items():
             topic = f"{MQTT_DISCOVERY}/{data['type']}/{cam.mac}/{entity}/config"
-            payload = base_payload | data["payload"]
-            payload[
-                "name"
-            ] = f"Wyze Cam {cam.nickname} {' '.join(entity.upper().split('_'))}"
-            payload["uniq_id"] = f"WYZE{cam.mac}{entity.upper()}"
+            if "availability_topic" not in data["payload"]:
+                data["payload"]["availability_topic"] = "wyzebridge/state"
+
+            payload = dict(
+                base_payload | data["payload"],
+                name=f"Wyze Cam {cam.nickname} {' '.join(entity.upper().split('_'))}",
+                uniq_id=f"WYZE{cam.mac}{entity.upper()}",
+            )
+
             msgs.append((topic, json.dumps(payload)))
+
     send_mqtt(msgs)
 
 
@@ -118,9 +78,11 @@ def mqtt_sub_topic(m_topics: list, callback) -> Optional[paho.mqtt.client.Client
 
     client.username_pw_set(MQTT_USER, MQTT_PASS or None)
     client.user_data_set(callback)
-    client.on_connect = lambda mq_client, *_: [
-        mq_client.subscribe(f"wyzebridge/{m_topic}") for m_topic in m_topics
-    ]
+    client.on_connect = lambda mq_client, *_: (
+        mq_client.publish("wyzebridge/state", "online"),
+        [mq_client.subscribe(f"wyzebridge/{m_topic}") for m_topic in m_topics],
+    )
+    client.will_set("wyzebridge/state", payload="offline", qos=1, retain=True)
     client.connect(MQTT_HOST, int(MQTT_PORT or 1883), 30)
     client.loop_start()
     return client
@@ -200,3 +162,102 @@ def _on_message(client, callback, msg):
     resp = callback(cam, topic, payload if include_payload else "")
     if resp.get("status") != "success":
         logger.info(f"[MQTT] {resp}")
+
+
+def get_entities(base_topic: str) -> dict:
+    return {
+        "snapshot": {
+            "type": "camera",
+            "payload": {
+                "availability_topic": f"{base_topic}state",
+                "payload_not_available": "stopped",
+                "topic": f"{base_topic}image",
+                "icon": "mdi:cctv",
+            },
+        },
+        "power": {
+            "type": "switch",
+            "payload": {
+                "command_topic": f"{base_topic}power/set",
+                "icon": "mdi:power-plug",
+            },
+        },
+        "ir": {
+            "type": "switch",
+            "payload": {
+                "state_topic": f"{base_topic}irled",
+                "command_topic": f"{base_topic}irled/set",
+                "payload_on": 1,
+                "payload_off": 2,
+                "icon": "mdi:lightbulb-night",
+            },
+        },
+        "night_vision": {
+            "type": "switch",
+            "payload": {
+                "state_topic": f"{base_topic}night_vision",
+                "command_topic": f"{base_topic}night_vision/set",
+                "payload_on": 3,
+                "payload_off": 2,
+                "icon": "mdi:weather-night",
+            },
+        },
+        "status_light": {
+            "type": "switch",
+            "payload": {
+                "state_topic": f"{base_topic}status_light",
+                "command_topic": f"{base_topic}status_light/set",
+                "payload_on": 1,
+                "payload_off": 2,
+                "icon": "mdi:led-on",
+                "entity_category": "diagnostic",
+            },
+        },
+        "bitrate": {
+            "type": "number",
+            "payload": {
+                "state_topic": f"{base_topic}bitrate",
+                "command_topic": f"{base_topic}bitrate/set",
+                "device_class": "data_rate",
+                "min": 1,
+                "max": 250,
+                "icon": "mdi:high-definition-box",
+                "entity_category": "diagnostic",
+            },
+        },
+        "fps": {
+            "type": "number",
+            "payload": {
+                "state_topic": f"{base_topic}fps",
+                "command_topic": f"{base_topic}fps/set",
+                "min": 1,
+                "max": 30,
+                "icon": "mdi:filmstrip",
+                "entity_category": "diagnostic",
+            },
+        },
+        "res": {
+            "type": "sensor",
+            "payload": {
+                "state_topic": f"{base_topic}res",
+                "icon": "mdi:image-size-select-large",
+                "entity_category": "diagnostic",
+            },
+        },
+        "signal": {
+            "type": "sensor",
+            "payload": {
+                "state_topic": f"{base_topic}wifi",
+                "icon": "mdi:wifi",
+                "entity_category": "diagnostic",
+            },
+        },
+        "audio": {
+            "type": "sensor",
+            "payload": {
+                "state_topic": f"{base_topic}audio",
+                "icon": "mdi:volume-high",
+                "entity_category": "diagnostic",
+            },
+        },
+    }